+++ lib/icons/index.ts
@@ -1,0 +1,8 @@
+export * from "./archive";
+export * from "./coingecko";
+export * from "./common";
+export * from "./cow-bucket";
+export * from "./dex-screener";
+export * from "./interface";
+export * from "./smoldapp";
+export * from "./zerion";
