+++ lib/icons/interface.ts
@@ -1,4 +1,4 @@
-import { TokenQuery } from "../types";
+import { TokenQuery } from "../schema";
 
 export interface IconFeed {
   name: string;
