+++ lib/prices/index.ts
@@ -4,3 +4,5 @@ export * from "./defilama";
 export * from "./dex-screener";
 export * from "./zerion";
 export * from "./interface";
+export * from "./revolver";
+export * from "./quorum";
