--- src/lib/icons/index.ts
@@ -1,4 +1,0 @@
-export * from "./archive";
-export * from "./cow-bucket";
-export * from "./interface";
-export * from "./zerion";
