--- src/lib/types.ts
@@ -1,6 +1,0 @@
-import { Address } from "viem";
-
-export interface TokenQuery {
-  address: Address;
-  chainId: number;
-}
