+++ tests/fixtures.ts
@@ -1,5 +1,5 @@
 import { Address } from "viem";
-import { TokenQuery } from "@/src/lib/types";
+import { TokenQuery } from "@/lib/schema";
 
 export const USDC_BASE = {
   address: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
