+++ tests/icons/zerion.spec.ts
@@ -1,5 +1,5 @@
 import { getZerionKey } from "@/src/app/config";
-import { ZerionIconFeed } from "@/src/lib/icons/zerion";
+import { ZerionIconFeed } from "@/lib/icons/zerion";
 import { TORN_MAINNET, TRUMP_BASE } from "../fixtures";
 // Fetching icon for 42161:0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2
 describe("Zerion Icon Archive", () => {
